Implement paging in google client
