+++ internal/infrastructure/google/user.go
@@ -1,0 +1,6 @@
+package google
+
+type User struct {
+	Id    string
+	email string
+}
